backend/requirements.txt
fastapi
uvicorn
google-genai
anthropic
python-dotenv
pydantic
openai